Separate s3 tests as per modules
